chỉnh sửa nho nhỏ

+++ database/migrations/2024_09_25_070943_create_products_table.php
@@ -20,8 +20,8 @@ return new class extends Migration
             $table->string('image');                        // Ảnh sản phẩm
             $table->string('SKU')->unique();                // Mã SKU
             $table->unsignedBigInteger('base_stock')->default(0);       // Số lượng tồn kho           
-            $table->decimal('price_regular');               // Giá thường
-            $table->decimal('price_sale')->nullable();      // Giá sale
+            $table->unsignedBigInteger('price_regular');               // Giá thường
+            $table->unsignedBigInteger('price_sale')->nullable();      // Giá sale
             $table->string('description')->nullable();      // Mô tả
             $table->integer('views')->default(0);    // Lượt xem
             $table->text('content')->nullable();            // Nội dung
